Split execution explanation by owner
- Separate function-table, label, and return-graph formatting into owner modules.
- Keep exact formatting tests beside each explanation responsibility.
- Register public explanation fixtures through the shared fixture harness.
