+++ tests/execution.rs
@@ -35,6 +35,23 @@ macro_rules! execution_error_cases {
     };
 }
 
+macro_rules! explain_cases {
+    ($($name:ident),+ $(,)?) => {
+        fixture_cases!(crate::run_explain_fixture, "explain"; $($name),+);
+    };
+}
+
+mod explain {
+    explain_cases!(
+        return_topology,
+        value_return_tables,
+        list_return_tables,
+        function_return_tables,
+        list_returning_function_tables,
+        return_table_group_order,
+    );
+}
+
 mod values {
     execution_cases!("values";
         integer_return,
@@ -956,197 +973,6 @@ fn constant_nested_list_rejects_mismatched_typed_ast_elements() {
     );
 }
 
-#[test]
-fn execution_plan_explain_formats_public_return_topology() {
-    let source = include_str!("fixtures/explain/return_topology.gleam");
-    let plan = execution_plan_for_explain(source);
-
-    assert_eq!(plan.explain().to_string(), expected_explanation(source));
-    assert_eq!(run_main(&plan), Ok(Value::Int(40.into())));
-}
-
-#[test]
-fn execution_plan_explain_formats_every_function_table_group() {
-    let fixtures = [
-        include_str!("fixtures/explain/value_return_tables.gleam"),
-        include_str!("fixtures/explain/list_return_tables.gleam"),
-        include_str!("fixtures/explain/function_return_tables.gleam"),
-        include_str!("fixtures/explain/list_returning_function_tables.gleam"),
-        include_str!("fixtures/explain/return_table_group_order.gleam"),
-    ];
-
-    for source in fixtures {
-        assert_eq!(
-            execution_plan_for_explain(source).explain().to_string(),
-            expected_explanation(source),
-        );
-    }
-}
-
-#[test]
-fn execution_plan_explain_labels_every_main_return_family() {
-    let cases = [
-        ("pub fn main() -> value { main() }", "never#0"),
-        ("pub fn main() { 1 }", "int#0"),
-        ("pub fn main() { 1.0 }", "float#0"),
-        ("pub fn main() { \"one\" }", "string#0"),
-        ("pub fn main() { <<1>> }", "bit_array#0"),
-        (
-            "pub fn main() -> UtfCodepoint { let assert <<value:utf8_codepoint>> = <<65>> value }",
-            "utf_codepoint#0",
-        ),
-        (
-            "pub type Boxed { Boxed(Int) } pub fn main() { Boxed(1) }",
-            "custom#0",
-        ),
-        ("pub fn main() { True }", "bool#0"),
-        ("pub fn main() { Nil }", "nil#0"),
-        ("pub fn main() { #(1) }", "tuple#0"),
-        ("pub fn main() -> List(value) { [] }", "list.parameter#0"),
-        (
-            "pub fn main() -> List(List(value)) { [[]] }",
-            "list.parameter_list#0",
-        ),
-        ("pub fn main() -> List(Int) { [] }", "list.int#0"),
-        ("pub fn main() -> List(String) { [] }", "list.string#0"),
-        ("pub fn main() -> List(BitArray) { [] }", "list.bit_array#0"),
-        (
-            "pub fn main() -> List(UtfCodepoint) { [] }",
-            "list.utf_codepoint#0",
-        ),
-        (
-            "pub type Boxed { Boxed(Int) } pub fn main() -> List(Boxed) { [] }",
-            "list.custom#0",
-        ),
-        ("pub fn main() -> List(Float) { [] }", "list.float#0"),
-        ("pub fn main() -> List(Bool) { [] }", "list.bool#0"),
-        ("pub fn main() -> List(Nil) { [] }", "list.nil#0"),
-        ("pub fn main() -> List(#(Int)) { [] }", "list.tuple#0"),
-        ("pub fn main() -> List(List(Int)) { [] }", "list.list#0"),
-        (
-            "pub fn main() -> List(fn() -> Int) { [] }",
-            "list.function#0",
-        ),
-        (
-            "pub fn main() -> fn(value) -> value { fn(value) { value } }",
-            "function.generic#0",
-        ),
-        (
-            "pub fn main() -> fn() -> value { fn() { panic } }",
-            "function.never#0",
-        ),
-        (
-            "pub fn main() -> fn() -> Int { fn() { 1 } }",
-            "function.int#0",
-        ),
-        (
-            "pub fn main() -> fn() -> Float { fn() { 1.0 } }",
-            "function.float#0",
-        ),
-        (
-            "pub fn main() -> fn() -> String { fn() { \"one\" } }",
-            "function.string#0",
-        ),
-        (
-            "pub fn main() -> fn() -> BitArray { fn() { <<1>> } }",
-            "function.bit_array#0",
-        ),
-        (
-            "pub fn main() -> fn() -> UtfCodepoint { fn() { panic } }",
-            "function.utf_codepoint#0",
-        ),
-        (
-            "pub type Boxed { Boxed(Int) } pub fn main() -> fn() -> Boxed { fn() { Boxed(1) } }",
-            "function.custom#0",
-        ),
-        (
-            "pub fn main() -> fn() -> Bool { fn() { True } }",
-            "function.bool#0",
-        ),
-        (
-            "pub fn main() -> fn() -> Nil { fn() { Nil } }",
-            "function.nil#0",
-        ),
-        (
-            "pub fn main() -> fn() -> #(Int) { fn() { #(1) } }",
-            "function.tuple#0",
-        ),
-        (
-            "pub fn main() -> fn() -> List(value) { fn() { [] } }",
-            "function.list.parameter#0",
-        ),
-        (
-            "pub fn main() -> fn() -> List(List(value)) { fn() { [[]] } }",
-            "function.list.parameter_list#0",
-        ),
-        (
-            "pub fn main() -> fn() -> List(Int) { fn() { [] } }",
-            "function.list.int#0",
-        ),
-        (
-            "pub fn main() -> fn() -> List(String) { fn() { [] } }",
-            "function.list.string#0",
-        ),
-        (
-            "pub fn main() -> fn() -> List(BitArray) { fn() { [] } }",
-            "function.list.bit_array#0",
-        ),
-        (
-            "pub fn main() -> fn() -> List(UtfCodepoint) { fn() { [] } }",
-            "function.list.utf_codepoint#0",
-        ),
-        (
-            "pub type Boxed { Boxed(Int) } pub fn main() -> fn() -> List(Boxed) { fn() { [] } }",
-            "function.list.custom#0",
-        ),
-        (
-            "pub fn main() -> fn() -> List(Float) { fn() { [] } }",
-            "function.list.float#0",
-        ),
-        (
-            "pub fn main() -> fn() -> List(Bool) { fn() { [] } }",
-            "function.list.bool#0",
-        ),
-        (
-            "pub fn main() -> fn() -> List(Nil) { fn() { [] } }",
-            "function.list.nil#0",
-        ),
-        (
-            "pub fn main() -> fn() -> List(#(Int)) { fn() { [] } }",
-            "function.list.tuple#0",
-        ),
-        (
-            "pub fn main() -> fn() -> List(List(Int)) { fn() { [] } }",
-            "function.list.list#0",
-        ),
-        (
-            "pub fn main() -> fn() -> List(fn() -> Int) { fn() { [] } }",
-            "function.list.function#0",
-        ),
-        (
-            "pub fn main() -> fn() -> fn() -> Int { fn() { fn() { 1 } } }",
-            "function.function#0",
-        ),
-    ];
-
-    for (source, expected) in cases {
-        let explanation = execution_plan_for_explain(source).explain().to_string();
-        let main = explanation
-            .lines()
-            .nth(1)
-            .expect("explanation should contain a main line");
-
-        assert_eq!(main, format!("main {expected}"));
-    }
-}
-
-fn execution_plan_for_explain(source: &str) -> geam::ExecutionPlan {
-    let module = compile_typed_module("main", "main.gleam", source)
-        .expect("explanation source should compile");
-    let module_plan = plan_module(module).expect("explanation source should plan");
-    geam::ExecutionPlan::from_module_plan(module_plan)
-}
-
 fn expected_explanation(source: &str) -> String {
     let (_, comments) = source
         .split_once("\n// geam:explain\n")
@@ -1162,6 +988,17 @@ fn expected_explanation(source: &str) -> String {
     }
 
     expected
+}
+
+fn run_explain_fixture(file_name: &str) {
+    let path = format!("tests/fixtures/{file_name}");
+    let source = std::fs::read_to_string(&path).expect("fixture should be readable");
+    let expected = expected_explanation(&source);
+    let module = compile_typed_module("main", path, &source).expect("fixture should compile");
+    let module_plan = plan_module(module).expect("fixture should plan");
+    let plan = geam::ExecutionPlan::from_module_plan(module_plan);
+
+    assert_eq!(plan.explain().to_string(), expected);
 }
 
 fn run_fixture(file_name: &str) {
